# codebase/pathos-backend/openapi.live.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PathOS Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audit/recent": {
      "get": {
        "summary": "Get Recent Audits",
        "operationId": "get_recent_audits_api_v1_audit_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Recent Audits Api V1 Audit Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit/{trace_id}": {
      "get": {
        "summary": "Get Audit",
        "operationId": "get_audit_api_v1_audit__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Audit Api V1 Audit  Trace Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Audit",
        "operationId": "delete_audit_api_v1_audit__trace_id__delete",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Audit Api V1 Audit  Trace Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/export/thread/{thread_id}": {
      "get": {
        "summary": "Export Thread",
        "operationId": "export_thread_api_v1_export_thread__thread_id__get",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Thread Api V1 Export Thread  Thread Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/audit/{trace_id}": {
      "get": {
        "summary": "Export Audit",
        "operationId": "export_audit_api_v1_export_audit__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Audit Api V1 Export Audit  Trace Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/recent": {
      "get": {
        "summary": "Export Recent",
        "operationId": "export_recent_api_v1_export_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Recent Api V1 Export Recent Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/saved-searches": {
      "post": {
        "summary": "Create Saved Search",
        "description": "Create deterministic saved search record.",
        "operationId": "create_saved_search_api_v1_saved_searches_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Saved Searches",
        "description": "List saved searches by recency.",
        "operationId": "list_saved_searches_api_v1_saved_searches_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedSearchOut"
                  },
                  "title": "Response List Saved Searches Api V1 Saved Searches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/saved-searches/{saved_search_id}": {
      "get": {
        "summary": "Get Saved Search",
        "description": "Fetch one saved search by identifier.",
        "operationId": "get_saved_search_api_v1_saved_searches__saved_search_id__get",
        "parameters": [
          {
            "name": "saved_search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saved Search Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Saved Search",
        "description": "Update one saved search record.",
        "operationId": "update_saved_search_api_v1_saved_searches__saved_search_id__put",
        "parameters": [
          {
            "name": "saved_search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saved Search Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSearchUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSearchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Saved Search",
        "description": "Delete saved search and related run/alert rows.",
        "operationId": "delete_saved_search_api_v1_saved_searches__saved_search_id__delete",
        "parameters": [
          {
            "name": "saved_search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Saved Search Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Saved Search Api V1 Saved Searches  Saved Search Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobScoringProfileV1": {
        "properties": {
          "preferred_series": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Preferred Series"
          },
          "preferred_grades": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreferredGrades"
              },
              {
                "type": "null"
              }
            ]
          },
          "preferred_locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Preferred Locations"
          },
          "remote_preference": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "no_preference",
                  "remote_only",
                  "onsite_only"
                ]
              },
              {
                "type": "boolean"
              }
            ],
            "title": "Remote Preference",
            "default": "no_preference"
          },
          "relocation_radius_miles": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Relocation Radius Miles"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Keywords"
          }
        },
        "type": "object",
        "title": "JobScoringProfileV1"
      },
      "JobSearchRequest": {
        "properties": {
          "keyword": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Keyword"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Only"
          },
          "grade_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 15.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Min"
          },
          "grade_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 15.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Max"
          },
          "series": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 25
              },
              {
                "type": "null"
              }
            ],
            "title": "Series"
          },
          "agency_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 25
              },
              {
                "type": "null"
              }
            ],
            "title": "Agency Codes"
          },
          "appointment_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Type"
          },
          "work_schedule": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Schedule"
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "keyword"
        ],
        "title": "JobSearchRequest",
        "description": "Validated search filters for deterministic USAJOBS query building.\n\nInputs:\n- Search keyword plus optional filters.\n\nOutputs:\n- Pydantic-normalized object with bounded fields.\n\nDeterministic assumptions:\n- Pagination is always explicit (`page`, `page_size`) to keep result windows predictable.\n- Grade range validation is stable and order-sensitive.\n\nError behavior:\n- Raises validation errors for empty/oversized inputs or invalid grade range."
      },
...
      "PreferredGrades": {
        "properties": {
          "min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 15.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 15.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          }
        },
        "type": "object",
        "title": "PreferredGrades"
      },
...
      "SavedSearchCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "query": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSearchRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobScoringProfileV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "ruleset_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Ruleset Version",
            "default": "job-scoring-v1"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSearchRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedSearchCreateRequest",
        "description": "Create payload for deterministic saved search rows."
      },
      "SavedSearchOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "query_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Query Payload"
          },
          "profile_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Payload"
          },
          "ruleset_version": {
            "type": "string",
            "title": "Ruleset Version"
          },
          "filters_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filters Json"
          },
          "is_enabled": {
            "type": "boolean",
            "title": "Is Enabled",
            "default": true
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "query_payload",
          "ruleset_version",
          "filters_json",
          "created_at",
          "updated_at"
        ],
        "title": "SavedSearchOut",
        "description": "API output contract for saved searches."
      },
      "SavedSearchUpdateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "query": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSearchRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobScoringProfileV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "ruleset_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Ruleset Version",
            "default": "job-scoring-v1"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobSearchRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedSearchUpdateRequest",
        "description": "Update payload for saved search rows."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}